requirements.txt
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.38
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1
python-telegram-bot==21.10